Rename `toast-helper.js` to `toast_helper.js`
