+++ app/javascript/controllers/input_child_info_controller.js
@@ -1,6 +1,6 @@
 import { Controller } from "@hotwired/stimulus";
 import { FetchRequest } from "@rails/request.js";
-import { showToast } from "helpers/toast-helper";
+import { showToast } from "helpers/toast_helper";
 
 export default class extends Controller {
   static targets = ["month", "year", "productCategory"];
