+++ app/javascript/helpers/toast_helper.js
