+++ app/javascript/turbo_streams/toast.js
@@ -1,4 +1,4 @@
-import { showToast } from "helpers/toast-helper";
+import { showToast } from "helpers/toast_helper";
 
 window.Turbo.StreamActions.toast = function() {
   const message = this.getAttribute("message")
